Splitting internal recommendation logic
